+++ packages/devtools_app_shared/CHANGELOG.md
@@ -4,6 +4,8 @@
 * Bump `vm_service` dependency to `^14.2.1`.
 * Add a `DTDManager.dispose` method.
 * Fix a race condition during service manager disconnect.
+* Add `IdeThemeQueryParams` extension type for parsing query params.
+* Add `EmbedMode` to enumerate the possible DevTools embedded states.
 
 ## 0.1.1
 * Update `package:dtd` to `^2.1.0`